workqueue: ensure attrs-changing be sequentially
Current modification to attrs via sysfs is not atomically.

Process A (change cpumask)	| Process B (change numa affinity)
wq_cpumask_store()		|
  wq_sysfs_prep_attrs()		|
				| apply_workqueue_attrs()
  apply_workqueue_attrs()	|

It results that the Process B's operation is totally reverted
without any notification.

This behavior is acceptable but it is sometimes unexpected.
Sequential model on non-performance-sensitive operations is more popular
and preferred. So this patch moves wq_sysfs_prep_attrs() into the protection
under wq_pool_mutex to ensure attrs-changing be sequentially.

This patch is also a preparation patch for next patch which change
the API of apply_workqueue_attrs().

Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>