workqueue: reuse the current per-node pwq when its attrs are unchanged
If the cpumask is changed, it is possible that only a part of the per-node
pwq is affected.  This can happen when the user changes the cpumask of
a workqueue or the low level cpumask.

So we try to reuse the current per-node pwq when its attrs are unchanged.

In wq_update_unbound_numa(), we had already made the current pwq be
reused when its attrs are unaffected, but we move the code of fetching
current pwq closer to the code of testing.

Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>