workqueue: reuse the current default pwq when its attrs unchanged
When apply_wqattrs_prepare() is called, it is possible that the default
pwq is unaffected.  It is always true that only the NUMA affinity is being
changed and sometimes true that the low level cpumask is being changed.

So we try to reuse the current default pwq when its attrs unchanged.

After this change, "ctx->dfl_pwq->refcnt++" could be dangerous
when ctx->dfl_pwq is a reusing pwq which may be receiving work items
or processing work items and hurts concurrency [get|put]_pwq(),
so we use get_pwq_unlocked() instead.

Signed-off-by: Lai Jiangshan <laijs@cn.fujitsu.com>